refactor: AutoTaggingService 책임 분리
