+++ backend/domain/src/main/java/com/techbloghub/domain/tagging/auto/port/RejectedItemPort.java
@@ -1,0 +1,9 @@
+package com.techbloghub.domain.tagging.auto.port;
+
+import com.techbloghub.domain.post.model.Post;
+import com.techbloghub.domain.tagging.auto.model.TaggingResult;
+
+public interface RejectedItemPort {
+
+    void saveRejectedItems(Post post, TaggingResult result);
+}
